requirements.txt
python-dotenv==1.0.1
llama-index-core==0.11.19
llama-parse==0.5.11
llama-index-readers-file==0.2.2
llama-index-embeddings-nvidia==0.2.3
llama-index-readers-opendal==0.0.1
pymupdf==1.24.10
llama-index-llms-nvidia==0.2.6
boto3==1.35.49
llama-index-vector-stores-milvus==0.2.7
openpyxl==3.1.5
torch==2.5.0
llama-index-embeddings-openai==0.2.5
llama-index-vector-stores-pinecone==0.2.1
pinecone-client>=3.0.0
googleapis-common-protos==1.65.0
pinecone[grpc]==5.1.0
fastapi==0.115.4
uvicorn==0.32.0
snowflake-connector-python==3.12.3
streamlit==1.39.0
python-multipart==0.0.17
streamlit-pdf-viewer==0.0.19
llama-index-callbacks-arize-phoenix==0.2.1
llama-index-multi-modal-llms-openai==0.2.3